requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Vector Search
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment if CUDA available and faiss-gpu installed

# UI
streamlit>=1.32.0

# Utilities
psutil>=5.9.0
tqdm>=4.66.0

# Optional: accelerated inference
# onnxruntime>=1.17.0
# optimum>=1.17.0
